Refine logging, adjust compiler options, and enhance module configuration
- Updated logging headers with conditional diagnostics for improved compatibility.
- Enhanced compiler options with stricter warnings and separate debug/release settings.
- Refactored target configurations, centralizing `configure_target_options` for consistency.
- Adjusted `node_count` return type in `Node` class from `unsigned int` to `long` for alignment.
- Improved `IRenderer` logging and type casting for clarity.
- Introduced new `spdlog_header.hpp` for centralized logger inclusions.

+++ include/bintjekit/configuration/node.hpp
@@ -37,7 +37,7 @@ namespace bnjkit {
             [[nodiscard]] nlohmann::json get_json() const;
             void reset_to_defaults();
             Node create_child(const nlohmann::json::json_pointer& key);
-            unsigned int node_count() const;
+            long node_count() const;
         protected:
             std::shared_ptr<spdlog::logger> m_logger;
             std::shared_ptr<nlohmann::json> m_json;
